refine ops and tf packages
