+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/tf/DataFlowOps.scala
@@ -13,11 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.nn.ops
+package com.intel.analytics.bigdl.nn.tf
 
 import java.util
 import java.util.concurrent.ConcurrentHashMap
 
+import com.intel.analytics.bigdl.nn.ops.Operation
 import com.intel.analytics.bigdl.nn.tf.WithoutInput
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
