+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/ArrayOps.scala
@@ -19,7 +19,7 @@ import java.nio.ByteOrder
 
 import com.intel.analytics.bigdl.Module
 import com.intel.analytics.bigdl.nn.Identity
-import com.intel.analytics.bigdl.nn.ops.{InvertPermutation => InvertPermutationOps,
+import com.intel.analytics.bigdl.nn.tf.{InvertPermutation => InvertPermutationOps,
   ConcatOffset => ConcatOffsetOps}
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
