+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/Div.scala
@@ -18,7 +18,7 @@ package com.intel.analytics.bigdl.utils.tf.loaders
 import java.nio.ByteOrder
 
 import com.intel.analytics.bigdl.Module
-import com.intel.analytics.bigdl.nn.ops.{RealDiv => RealDivOp}
+import com.intel.analytics.bigdl.nn.CDivTable
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
 import org.tensorflow.framework.NodeDef
@@ -28,6 +28,6 @@ import scala.reflect.ClassTag
 class Div extends TensorflowOpsLoader {
   override def build[T: ClassTag](nodeDef: NodeDef, byteOrder: ByteOrder,
                                   context: Context[T])(implicit ev: TensorNumeric[T]): Module[T] = {
-    RealDivOp()
+    CDivTable().asInstanceOf[Module[T]]
   }
 }
