+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/ExpandDims.scala
@@ -18,9 +18,8 @@ package com.intel.analytics.bigdl.utils.tf.loaders
 import java.nio.ByteOrder
 
 import com.intel.analytics.bigdl.Module
-import com.intel.analytics.bigdl.nn.Identity
+import com.intel.analytics.bigdl.nn.Unsqueeze
 import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity}
-import com.intel.analytics.bigdl.nn.ops.ExpandDims
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
@@ -41,6 +40,6 @@ class ExpandDims extends TensorflowOpsLoader {
 class ExpandDimsLoadTF[T: ClassTag]()(implicit ev: TensorNumeric[T]) extends Adapter[T](Array(2)) {
   override def build(tensorArrays: Array[Tensor[_]]): AbstractModule[Activity, Activity, T] = {
     val axis = tensorArrays(0).asInstanceOf[Tensor[Int]].value()
-    ExpandDims[T](if (axis < 0) axis + 1 else axis + 1)
+    Unsqueeze[T](if (axis < 0) axis + 1 else axis + 1)
   }
 }
