+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/ParseExample.scala
@@ -18,7 +18,7 @@ package com.intel.analytics.bigdl.utils.tf.loaders
 import java.nio.ByteOrder
 
 import com.intel.analytics.bigdl.Module
-import com.intel.analytics.bigdl.nn.ops.{ParseExample => ParseExampleOperation}
+import com.intel.analytics.bigdl.nn.tf.{ParseExample => ParseExampleOperation}
 import com.intel.analytics.bigdl.tensor._
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
