+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/GraphSpec.scala
@@ -23,8 +23,8 @@ import com.intel.analytics.bigdl.models.lenet.LeNet5
 import com.intel.analytics.bigdl.models.vgg.{VggForCifar10, Vgg_16, Vgg_19}
 import com.intel.analytics.bigdl.nn.Graph.ModuleNode
 import com.intel.analytics.bigdl.nn.abstractnn.EmptyGradInput
-import com.intel.analytics.bigdl.nn.ops.{Ceil, ControlNodes, Less}
-import com.intel.analytics.bigdl.nn.tf.Const
+import com.intel.analytics.bigdl.nn.ops.{Ceil, Less}
+import com.intel.analytics.bigdl.nn.tf.{Const, ControlNodes}
 import com.intel.analytics.bigdl.numeric.NumericFloat
 import com.intel.analytics.bigdl.utils.RandomGenerator._
 import com.intel.analytics.bigdl.tensor.Tensor
