+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/ops/ExpandDimsSpec.scala
@@ -16,7 +16,7 @@
 package com.intel.analytics.bigdl.nn.ops
 
 import com.intel.analytics.bigdl.tensor.Tensor
-import com.intel.analytics.bigdl.utils.T
+import com.intel.analytics.bigdl.nn.{Unsqueeze => ExpandDims}
 import org.scalatest.{FlatSpec, Matchers}
 
 class ExpandDimsSpec extends FlatSpec with Matchers {
