+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/tf/ConstSpec.scala
@@ -28,31 +28,10 @@ class ConstSpec extends FlatSpec with Matchers {
     layer.forward(input) should be(value)
   }
 
-  "Const backward tensor" should "be correct" in {
-    val value = Tensor(2, 3).rand()
-    val layer = Const(value)
-    val input = Tensor(4, 5).rand()
-    val gradOutput = Tensor(2, 3).rand()
-    layer.forward(input) should be(value)
-    val grad = layer.backward(input, gradOutput).toTensor
-    grad should be(Tensor(4, 5).zero())
-  }
-
   "Const forward tensors" should "be correct" in {
     val value = Tensor(2, 3).rand()
     val layer = Const(value)
     val input = T(Tensor(4, 5).rand(), Tensor(3, 4).rand())
     layer.forward(input) should be(value)
-  }
-
-  "Const backward tensor" should "be correct when input is tensors" in {
-    val value = Tensor(2, 3).rand()
-    val layer = Const(value)
-    val input = T(Tensor(4, 5).rand(), Tensor(3, 4).rand())
-    val gradOutput = Tensor(2, 3).rand()
-    layer.forward(input) should be(value)
-    val grad = layer.backward(input, gradOutput).toTable
-    grad[Tensor[Float]](1) should be(Tensor(4, 5).zero())
-    grad[Tensor[Float]](2) should be(Tensor(3, 4).zero())
   }
 }
