+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/tf/ParseExampleSpec.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.nn.ops
+package com.intel.analytics.bigdl.nn.tf
 
 import com.intel.analytics.bigdl.tensor.{FloatType, LongType, StringType, Tensor}
 import com.google.protobuf.{ByteString, CodedOutputStream}
